# PruebaTecnica/src/main/java/com/unicomer/dto/ApiData.java
package com.unicomer.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ApiData {
	private String name;
	private Object data;
}