# api-docs.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"OpenAPI definition",
      "version":"v0"
   },
   "servers":[
      {
         "url":"http://localhost:8080",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/v1/customers/{id}":{
         "get":{
            "tags":[
               "customer-controller"
            ],
            "operationId":"findById",
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int64"
                  }
               }
            ],
            "responses":{
               "500":{
                  "description":"Internal Server Error",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               },
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               }
            }
         },
         "put":{
            "tags":[
               "customer-controller"
            ],
            "operationId":"updateCustomer",
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int64"
                  }
               }
            ],
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/Customer"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "500":{
                  "description":"Internal Server Error",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               },
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               }
            }
         }
      },
      "/v1/customers":{
         "get":{
            "tags":[
               "customer-controller"
            ],
            "operationId":"findAll",
            "responses":{
               "500":{
                  "description":"Internal Server Error",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               },
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "customer-controller"
            ],
            "operationId":"createCustomer",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/Customer"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "500":{
                  "description":"Internal Server Error",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               },
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ApiResult"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "ApiData":{
            "type":"object",
            "properties":{
               "name":{
                  "type":"string"
               },
               "data":{
                  "type":"object"
               }
            }
         },
         "ApiResult":{
            "type":"object",
            "properties":{
               "code":{
                  "type":"string"
               },
               "message":{
                  "type":"string"
               },
               "timestamp":{
                  "type":"string",
                  "format":"date-time"
               },
               "data":{
                  "$ref":"#/components/schemas/ApiData"
               }
            }
         },
         "Customer":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"integer",
                  "format":"int64"
               },
               "customerId":{
                  "type":"string"
               },
               "firstName":{
                  "type":"string"
               },
               "lastName":{
                  "type":"string"
               },
               "birthday":{
                  "type":"string"
               },
               "gender":{
                  "type":"string"
               },
               "cellPhone":{
                  "type":"string"
               },
               "homePhone":{
                  "type":"string"
               },
               "homeAddress":{
                  "type":"string"
               },
               "profession":{
                  "type":"string"
               },
               "income":{
                  "type":"string"
               }
            }
         }
      }
   }
}